Separate helper concerns
Don't overload the avatar helper with mentions related helpers.
