+++ app/helpers/comments_helper.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module CommentsHelper
+  def comment_formatter(comment)
+    mentions_formatter(simple_format(h(comment))).html_safe
+  end
+end
